Commented away the Specular color because it is unused and was spamming error messages. Doing some ground work for the future lobby. Setting up the logic on how the team picking should be handled etc.

+++ DV1573---UD1448/Networking/NetworkGlobals.h
@@ -7,7 +7,7 @@ namespace NetGlobals {
 	constexpr unsigned short MaximumIncomingConnections = 12;
 	constexpr unsigned short ServerPort = 42405;
 	constexpr int tickRate = 128;
-	constexpr float threadSleepTime = (1.0f / static_cast<float>(tickRate)) * 1000.0f;
+	constexpr unsigned int threadSleepTime = (1.0f / static_cast<float>(tickRate)) * 1000u;
 
 	enum Teams {
 		RedTeam,
